backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic[email]==2.10.4
python-multipart==0.0.19
boto3==1.36.4
Pillow==11.1.0
stripe==11.4.1
gunicorn==23.0.0
asyncpg==0.30.0
oracledb==2.5.1
oci==2.168.0
